requirements.txt
python-dotenv
gspread
oauth2client
openai
streamlit